init lab3

+++ util.c
@@ -22,7 +22,7 @@ string String(char *s)
 }
 
 U_boolList U_BoolList(bool head, U_boolList tail)
-{ U_boolList list = checked_malloc(sizeof(*list));
+{ U_boolList list = checked_malloc((int)sizeof(*list));
   list->head = head;
   list->tail = tail;
   return list;
